chore: 스토리 카드 폴더 위치 변경[BETA-100]
