+++ src/app/storybook/page.tsx
@@ -1,6 +1,6 @@
-import StoryCardGrid from "@/entities/story/ui/story-card/StoryCardContent";
-import data from "@/shared/ui/storycard/mock.json";
-import type { StoryEntity } from "@/shared/types/story.types";
+import StoryCardGrid from "@/widgets/storycard-container/StoryCardContainer";
+import data from "@/shared/mock/story.json";
+import type { StoryEntity } from "@/entities/story/story.types";
 
 const stories: StoryEntity[] = Array(9).fill(data).map((item, idx) => ({
   ...item,
