+++ src/entities/story/story.types.ts
@@ -1,0 +1,10 @@
+interface StoryEntity {
+  id: number;          
+  title: string;          
+  description?: string;   
+  category: string;       
+  icon: string; 
+  articleIds: number[]
+}
+
+export type { StoryEntity };
