+++ src/entities/story/ui/story-card/StoryCard.css.ts
