+++ src/entities/story/ui/story-card/storyIconMap.tsx
@@ -1,5 +1,6 @@
 import ChartIcon from "@/shared/assets/icons/common/chart.svg"
 
-export const mockStoryIconMap: Record<string, React.ReactNode> = {
+export const storyIconMap: Record<string, React.ReactNode> = {
   chart: <ChartIcon/>,
 };
+
